Add package for klausuromat Update readme Move content of gen.py into package __init__.py
