+++ klausuromat/include/generator.coffee
