+++ klausuromat/include/jquery.snippet.js
