+++ klausuromat/operations/eq_.py
