+++ klausuromat/operations/lshift_.py
