+++ klausuromat/operations/result.py
